pyproject.toml
[tool.poetry]
name = "swiftrank"
version = "1.3.0"
description = "Compact, ultra-fast SoTA reranker enhancing retrieval pipelines and terminal applications."
authors = ["Harsh Verma <synacktra.work@gmail.com>"]
license = "Apache Software License (Apache 2.0)"
readme = "readme.md"

[tool.poetry.dependencies]
python = ">=3.8, <4.0"
tokenizers = "^0.15.0"
onnxruntime = "1.16.3"
numpy = ">=1.24.4"
requests = "2.31.0"
rich = "13.7.0"
tqdm = "4.66.1"
cyclopts = "2.1.2"
pyyaml = "6.0.1"
orjson = "3.9.10"
pydantic = "2.6.4"
fastapi = "0.110.1"
uvicorn = "0.29.0"

[tool.poetry.scripts]
swiftrank = "swiftrank.interface.cli:app.meta"
srank = "swiftrank.interface.cli:app.meta"

[tool.poetry.group.dev.dependencies]
pytest = "8.1.1"
requests = "2.31.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
